Move key to the crate root
Done as part of flattening the `util` module.

Move the `key` module to the crate root, deprecated the module in `util`
and fix all paths in the codebase to use the new path.
