+++ src/blockdata/script.rs
@@ -33,7 +33,7 @@ use crate::policy::DUST_RELAY_TX_FEE;
 #[cfg(feature="bitcoinconsensus")] use core::convert::From;
 use crate::OutPoint;
 
-use crate::util::key::PublicKey;
+use crate::key::PublicKey;
 use crate::address::WitnessVersion;
 use crate::util::taproot::{LeafVersion, TapBranchHash, TapLeafHash};
 use secp256k1::{Secp256k1, Verification, XOnlyPublicKey};
@@ -1117,7 +1117,6 @@ mod test {
     use crate::hashes::hex::{FromHex, ToHex};
     use crate::consensus::encode::{deserialize, serialize};
     use crate::blockdata::opcodes;
-    use crate::util::key::PublicKey;
     use crate::util::psbt::serialize::Serialize;
     use crate::internal_macros::hex_script;
