+++ src/lib.rs
@@ -85,6 +85,7 @@ pub mod blockdata;
 pub mod consensus;
 pub mod error;
 pub mod hash_types;
+pub mod key;
 pub mod policy;
 pub mod util;
 
@@ -105,10 +106,10 @@ pub use crate::blockdata::witness::{self, Witness};
 pub use crate::blockdata::{constants, opcodes};
 pub use crate::consensus::encode::VarInt;
 pub use crate::hash_types::*;
+pub use crate::key::{KeyPair, PrivateKey, PublicKey, XOnlyPublicKey};
 pub use crate::network::constants::Network;
 pub use crate::util::amount::{Amount, Denomination, SignedAmount};
 pub use crate::util::ecdsa::{self, EcdsaSig, EcdsaSigError};
-pub use crate::util::key::{KeyPair, PrivateKey, PublicKey, XOnlyPublicKey};
 pub use crate::util::merkleblock::MerkleBlock;
 pub use crate::util::schnorr::{self, SchnorrSig, SchnorrSigError};
 pub use crate::util::sighash::{
